+++ libsolidity/interface/AssemblyStack.h
@@ -24,6 +24,9 @@
 #include <liblangutil/ErrorReporter.h>
 #include <liblangutil/EVMVersion.h>
 
+#include <libyul/Object.h>
+#include <libyul/ObjectParser.h>
+
 #include <libevmasm/LinkerObject.h>
 
 #include <string>
@@ -32,12 +35,6 @@
 namespace langutil
 {
 class Scanner;
-}
-
-namespace yul
-{
-struct AsmAnalysisInfo;
-struct Block;
 }
 
 namespace dev
@@ -72,10 +69,6 @@ public:
 	/// Multiple calls overwrite the previous state.
 	bool parseAndAnalyze(std::string const& _sourceName, std::string const& _source);
 
-	/// Runs analysis step on the supplied block, returns false if input cannot be assembled.
-	/// Multiple calls overwrite the previous state.
-	bool analyze(yul::Block const& _block, langutil::Scanner const* _scanner = nullptr);
-
 	/// Run the assembly step (should only be called after parseAndAnalyze).
 	MachineAssemblyObject assemble(Machine _machine) const;
 
@@ -94,8 +87,7 @@ private:
 	std::shared_ptr<langutil::Scanner> m_scanner;
 
 	bool m_analysisSuccessful = false;
-	std::shared_ptr<yul::Block> m_parserResult;
-	std::shared_ptr<yul::AsmAnalysisInfo> m_analysisInfo;
+	std::shared_ptr<yul::Object> m_parserResult;
 	langutil::ErrorList m_errors;
 	langutil::ErrorReporter m_errorReporter;
 };